[BUG]: wrap `handleChange` callback with Ember.run
The Ember run-loop efficiently manages jobs that are created during the
life of an application.  However, async events that are triggered by
non-Ember code can escape the run-loop.  This results in inefficiencies
and unexpected behavior.

Ember recommends that all non-Ember async callbacks be wrapped with
`Ember.run`, which causes the code to be run inside a run-loop.  This
commit wraps `WrappedComponent.handleChange`, which is the callback
provided to `redux.subscribe`.