Use Ember.Object.setProperties to update props
Create private properties that are updated with Ember.Object.set, and
then create public read-only aliases as specified by the user's
`mapStateToComputed` function provided to `connect`.  This replaces the
existing logic that accomplished this by directly manipulating public
properties and firing `notifyPropertyChange`.  Ember.Object.set is much
simpler and allows Ember to take charge of optimizing execution.

This commit also refactors the style and spirit of the connect function.
Rather than using functions behind-the-scenes for computing state, this
commit adds public methods to the component being wrapped.  This is more
in line with traditional Ember addons, and allows the user to customize
behavior if that was ever desired.

Specifically, we add (or modify):

    - WrappedComponent.getPrivateKey
    - WrappedComponent.transformPropsToPrivate
    - WrappedComponent.updateProps

Finally, several minor ES6 syntax changes are included.