feat(connect): pass attrs to stateToComputed
This allows users to reference attrs that are passed to their component
when preparing the returned properties.

    {{my-component myName="John"}}

    stateToComputed = (state, attrs) => {
      return { user: getUserByName(state, attrs.myName) };
    }

This is analogous to the same API in react-redux's mapStateToProps
function.  react-redux also does the same thing for the dispatch
function, however that is more complicated in Ember because actions are
not easily redefined—so changes to attrs would not be reflected as the
user would expect.  Therefore there are no changes to dispatchToActions
in this commit.

Closes #54