chore(build): drop browserify
This commit removes ember-browserify as a dependency, relying instead on
the UMD distributions of redux and redux-thunk.

The dist files are brought in via treeForVendor, and the resulting files
are imported as named AMD modules via ember-cli's new transformation
option for app.import available in ember-cli >=2.9.

The build detects if it is for the addon itself, and adds redux-saga in
that case for the tests that require it.

More information on ember-cli app.import transformations:

    https://github.com/ember-cli/ember-cli/pull/5976

It should be noted that while this does not change compatability with
Ember itself, users must be using ember-cli >= 2.9.

See #28