chore(docs): use /docs on master + GitHub pages
This leverages the latest and greatest from GitHub which allows you to
create Markdown files in a /docs directory on master + all the usual
Jekyll files/configs.

Hosting can be configured in the project settings, and Jekyll is
configured in the usual _config.yml file.

See the GitHub blog post for more information:

    https://github.com/blog/2289-publishing-with-github-pages-now-as-easy-as-1-2-3

Ref #61